# Designs/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "IteraWebApi",
    "version": "1.0"
  },
  "paths": {
...
    "/api/UserRoles/GetUserRolesAsync": {
      "post": {
        "tags": [
          "UserRoles"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetUserRolesQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetUserRolesQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetUserRolesQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/UserRoles/UserHasRoleAsync": {
      "post": {
        "tags": [
          "UserRoles"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/UserRoles/AssignRoleToUserAsync": {
      "post": {
        "tags": [
          "UserRoles"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleToUserCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleToUserCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleToUserCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/UserRoles/RemoveRoleFromUserAsync": {
      "post": {
        "tags": [
          "UserRoles"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveRoleFromUserCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveRoleFromUserCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveRoleFromUserCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/UserRoles/GetUsersInRoleAsync": {
      "post": {
        "tags": [
          "UserRoles"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetUsersInRoleQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetUsersInRoleQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetUsersInRoleQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AssignRoleToUserCommand": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "nullable": true
          },
          "roleId": {
            "type": "string",
            "nullable": true
          },
          "assignedBy": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GetUserRolesQuery": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetUsersInRoleQuery": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RemoveRoleFromUserCommand": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "nullable": true
          },
          "roleId": {
            "type": "string",
            "nullable": true
          },
          "removedBy": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}